refactor(InputView): InputValidator 추가 및 inputGameStatus 메소드 리팩토링
