+++ src/main/java/baseball/domain/dto/GameResult.java
@@ -1,0 +1,4 @@
+package baseball.domain;
+
+public record GameResult(int ball, int strike, GameStatus gameStatus) {
+}
