+++ src/main/java/baseball/refactoring/GameStatus.java
